requirements.txt
anthropic>=0.25.0
kfp>=2.7.0
kfp-server-api>=2.0.0
mlflow>=2.12.0
kubernetes>=29.0.0
PyGithub>=2.3.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
